requirements.txt
numpy
pandas
scikit-learn
scikit-image
mahotas
bitdesc
opencv-python
matplotlib
scipy
streamlit
joblib